nivola-docs-demo/requirements.txt
chromadb>=0.4.4
openai>=1.7.2
python-dotenv